pyproject.toml
[project]
name = "heat-mortality-italy"
version = "0.1.0"
description = "Extreme heat and regional mortality in Italy (2012-2022): the role of social vulnerability"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Data manipulation
    "pandas>=2.2",
    "numpy>=1.26",
    "openpyxl>=3.1",

    # Climate data
    "xarray>=2024.1",
    "netcdf4>=1.6",
    "cdsapi>=0.7",
    "cfgrib>=0.9",
    "dask>=2024.1",

    # Geospatial
    "geopandas>=0.14",
    "shapely>=2.0",
    "regionmask>=0.12",
    "pyproj>=3.6",

    # Statistical analysis
    "linearmodels>=6.0",
    "statsmodels>=0.14",
    "scipy>=1.12",
    "scikit-learn>=1.4",

    # Visualization
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.18",

    # Utilities
    "pyyaml>=6.0",
    "requests>=2.31",
    "tqdm>=4.66",
    "click>=8.1",
    "loguru>=0.7",
]

[project.scripts]
run-pipeline = "scripts.run_pipeline:main"
download-data = "scripts.download_data:main"
run-analysis = "scripts.run_analysis:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "pytest>=9.0.3",
]
